simplelink: dpl: cc13x2_cc26x2: improve ClockP accuracy
The existing implementation of ClockP was based on k_timer, which has
millisecond resolution, thus limiting ClockP's accuracy. To improve on
this, we take the nortos implementation of ClockP and TimerP, and
modify TimerP to directly call the Zephyr's timeout API. The latter has
tick-based accuracy which helps ClockP to schedule more precise timing.

The implementation maintains its own list of ClockP deadlines, and
adds the soonest one to z_timeout. When it expires, it would
call ClockP_doTick to execute code for any expired deadline. To kick
things off, we schedule an initial timeout in Zephyr by calling
ClockP_doTick once in ClockP_startup.

Signed-off-by: Vincent Wan <vincent.wan@linaro.org>